requirements.txt
llama-cpp-python
gradio
numpy<2.0.0
